Automatically create child documents in SaveDocument mutation
When a form contains FormQuestions, those _always_ need to have a sub
document attached - otherwise, the form can never be complete.
The creation of subforms could be done in the frontend (in fact that was
my first approach), but that is tedious, requires many requests, and
also feels somewhat redundant.

This adapts the "SaveDocument" mutation such that it recursively creates
subdocuments for all FormQuestions and attaches them to the newly
created main document as answers.