Add data_sources for dynamic options in select fields (#377)
* Add data_sources for dynamic options in select fields

This commit implements the `data_sources` extension point and the new
question-types `DynamicChoiseQuestion` and
`DynamicMultipleChoiceQuestion`.

There are two queries:
1. allDataSources --> Get all configured sources
2. dataSource --> get data from source

Closes #320

* Some cleanup after code review

* Do not auto-discover data_sources

This commit removes auto-discovery of data sources.
Instead the setting `DATA_SOURCE_CLASSES` is used.

* Do not create option objects for dynamic choice questions

This commit removes the creation of option-objects. Instead the choices
are refetched every time the question is accessed.

* Some cleanup

* fix default value of data_source.timeout
* remove oprions from dynamic question serializers

* Use multilingual values for DataSource.info and Data.label

Also updated the README accordingly.