New filter on MetaFilterSet: metaValue (#393)
This new filter is able to filter any node that has a meta field by the
meta field's value, as follows:

   query foo {
     allDocuments (metaValue: {key: "foo", value: "hello world"}) {
       ...
       }
   }

This will return all documents whose meta field has a key named "foo"
with the value "hello world".
An optional parameter can be given to the filter to configure the lookup
type:

   metaValue: {key: "foo", value: "hello", lookup:STARTSWITH}

This would change the filter to startswith (prefix search). Also
available lookup methods are EXACT (default if parameter is left out,
and CONTAINS (which does a case-insensitive lookup)).