Implement flexible format validators
This commit adds flexible format validators.

There is one new query: `all_format_validators` that enables querying
all available format validators.

A text- or textarea-question can have one or multiple format_validators
assigned.

When querying a question, there is a new `FormatValidatorsConnection`,
containing all FormatValidator-objects assigned on the question.

For this there is a new `QuestionValidator` that ensures, referenced
format_validators exist.

`translate_value()` has been moved to `core.utils`, as this is used in the
`data_source`- and `form`-app.

There is a new extension-point `format_validators.py`.

When adding the tests for this, it became obvious, that there was something
wrong with the parameters for `DataSource`. As I was touching those tests
anyway, I fixed this as well.

Added commented out extension volume to docker-compose.yml. Also added
forgotten data_sources extension volume.

Additionally this commit also cleans up the snapshots for document and
question as there were quite a lot old unused snapshots.

Closes #360