Fix returned types
Graphenes auto type detection doesn't work reliably with our different
Question and Answer types.

This commit adds a subclass of our subclass of DjangoObjectType with
support for resolving Question and Answer types.

Additionally the type is explicitly set for `parent_answers` of `Document`
and `answer` on `File`.

Also updated the `AnswerFactory` to correctly set float and int values for
their respective Answers.

Closes #432