Fallback value for all empty answer types
When evaluating a form's validity (is_required, is_hidden), we need
access to some value of all question types. In some cases, the answer
object's "value" field is None because the value is somewhere else, and
in other cases it's None because the answer really doesn't contain a
value.

This ensures we get a value in any case.