Major cleanup of validation procdure
There were several problems when validating the whole form.

First, the answer structure was not complete in some cases, meaning that
the "root" lookup never really worked: we always only had one level of
"parent", even in a nested form.

Then, required must not be evaluated if the question is in a subform
whose question is not marked as required. Some test assumed this
wrongly.

In a previous PR, we changed the "empty" result of an unanswered
question of various types to be an empty list to avoid failing JEXL
expressions (mainly the listy-types like multiple choice and form).
However when checking for emptyness (if the question was required), the
empty list was considered "present". This caused some validations to
assume "OK" even if the answer was empty.

Note: Some TODOs remain as there is still quite a bit of optimisation
potential in that bit of code