feat(caluma): Remove FormAnswer type
There was a conceptual misunderstanding that led us to implement
nested answers. Which in turn created performance and complexity
issues.

This commit removes the FormAnswer type. Additionally also all
implementations of recursive lookups and form path handling (that only existed
as a result of this misconception) are removed.

Closes #504

BREAKING CHANGE:
 - The FormAnswer type is removed and answers to a question
   in a sub-form do not have their own document.
 - In order to access parent questions from within a table-sub-question
   JEXL, there is no need to prefix the question_slug with `parent.`
   anymore.