Performance improvements
Profiling the validation process showed that a lot of time is spent in
parsing (not even executing!) the JEXL expressions.

Caching the parsed expressions (and shortcutting simple "true" / "false"
expressions) improved the validation time for a large-ish document
structure from roughly 46s on my machine down to 23s.

Of the remaining 23s, 17s are spent fetching data from the DB, which
will be sped up when the "flattening" of the document structure is
merged