Use django_simple_history
This commit adds `django_simple_history` in order to record changes to the
records.

This is implemented in the `BaseModel`, so the history is available in
all models.

Via the envvar `SIMPLE_HISTORY_ACTIVE` the history can be disabled. Switching
this setting results in two migrations.

Changes due to the [limitations](https://django-simple-history.readthedocs.io/en/2.7.2/common_issues.html) when using `django_simple_history`:

* Replaced `bulk_create` with `bulk_create_with_history` where used.
* Replaced `queryset.update()` with a for loop where used.

Other:

* Fixed a race condition in `caluma.workflow.tests.test_case.test_status_filter()`.

Closes #516