feat(filter): hasAnswer: new INTERSECTS lookup
The intersects lookup expects the search value to be a list of strings,
which is then mached against the value of the answer.

For example, given you've got a question of type "choice", with the
value "foo":

Then the following query would return the corresponding document:

        query asdf ($hasAnswer: [{question: "some-slug", lookup: INTERSECTS, value: ["foo", "bar"]}]) {
          allDocuments(hasAnswer: $hasAnswer) {
            edges {
              node {
                id
              }
            }
          }
        }

With multiple choice fields, it is similar - if any of the selected
choices matches any of the given search values, the corresponding
document is returned.