Do not swallow question is missing exception
As Caluma validates jexl when saving it technically the only exception
which can happen here is that an answer is missing.
Hence removing whole block.

It might make sense to have a possibility to validate all jexl
of a form to avoid Runtime error but this should happen outside
the document validation.