feat(filters): Text search across multiple answers
This introduces a filter named `searchAnswers` that allows you to search
text within documents/cases across multiple answers.

The way it works is you pass a list of questions to search, and a search
string. The search string is split by words. Each word must then be
found in at least one answer for a document/case to show up.

You may optionally provide a lookup type: `STARTSWITH`, `TEXT`,
`CONTAINS`.

Note that depending on the database, those may or may not perform well,
as we do not at this point in time preemptively optimize by creating
indexes etc.