chore(uwsgi): sane defaults for uWSGI
The defaults of uWSGI are not good. These are the settings researched by
Peter Sperl at bloomberg and @winpat.

I checked that uWSGI accepts overriding these settings by
environment-variables and composed gist for the admins wanting to change
our defaults. https://git.io/JemA2

It is gist because it is interesting for other projects and caluma doesn't
have a place for this kind of documentation, yet. It would have been out of
place in the README which is about caluma not uWSGI.