feat(filters): factory for unionized filters
Introduce `FilterCollectionFactory` and `CollectionFilterSetFactory`.
Those convert a regular FilterSet into a single filter that accepts
a list of filters as parameters. Additionally, they also allow
inversion of filters.

The `FilterCollectionFactory` takes a `FilterSet` and returns a `Filter`
class that contains the filterset's type as an inner type, while adding the
`invert` option. The resulting filter applies all the given filters
in turn to the queryset.

The `CollectionFilterSetFactory` takes this a bit further and returns
a `FilterSet` class that contains a single `filter` attribute, containing
the result of the `FilterCollectionFactory`.