refactor(validators): collect required questions
We iterate over the questions of the form via objects instead of the
queryset values() dict as it doesn't provide any additonal value in
this context.

Additionally (and the actual intent of this change), we return a list
of required questions. This will be used in the caller to decide which
answers to validate (hidden questions should not be validated)