refactor: move is_completed local function to method
The `is_completed(task)` function within the `update()` method of
the CompleteWorkItemSerializer was confusing to read, as it wasn't
clear about it's purpose (telling the caller whether the process
can continue if all siblings are completed).

Moving to a method for future subclassing, for example in the `Skip`
serializer coming up next.