fix(validation): correctly handle is_hidden regarding dependencies
When a question's `is_hidden` expression references other questions,
we need to check their visibility as well: if they're all hidden,
the user is unable to fill them in, making it impossible to decide
the first question's visibility. In those cases, we make the dependent
question hidden as well.

Following this logic, we need to apply the same principle to `is_required`
as well: if `is_required` references questions that are all hidden,
we cannot depend on their value to make the decision, so we assume
the question to be not required.