refactor: requiredness/visiblility validation
Thre requiredness and visibility was rather complex and error-prone.
The fact that we need to fully evaluate visibility recursively didn't help
either, so the ful visibility/requiredness analysis had to be refactored.

With the use of a context that is passed down during the process, some SQL
queries can be avoided while simplifying some things as well.