feat(form): allow Answers with empty value
This commit adds the possibility to save/query Answers with `null`
values.

It also adds a new lookup mode `ISNULL` to the `HasAnswerFilter`. If
this lookup mode is used, the provided `value` (if any) is ignored.

Since graphql-core doesn't support `null` literals yet, the way for setting
the value of an Answer to `None` is to omit the `value` field
altogether.

Closes #790

BREAKING CHANGE: `Answers` are now allowed to have `null` values.