fix(validation): correctly handle indirectly hidden questions
Questions can not only be hidden by evaluating their is_hidden expression
but also when they're part of a subform, where the containing form question
is hidden. We need to check for that case as well, which didn't happen till now.

This is a small refactoring of the validation process that separates checking question
visibility checks from the requiredness checks, enabling us to get a list of visible
questions before starting to evaluate requiredness (which depends on having such a
list of visible questions available to function correctly)