fix(pypi): fix loading of caluma settings
Because of the * import in `settings.__init__`, it's was not possible to
just load the caluma specific settings, without implicitly also loading
the django project settings.

This commit fixes that. The downside of this fix is, that the settings
module is not called `caluma.settings` anymore, but
`caluma.settings.django`.