chore(urls): move view and add urls.py to core
This commit moves the graphql view to the core app and renames it to
`CalumaGraphQLView`. Additionally it adds a `caluma_core.urls` that gets
included in the main `urls.py`.

This means third-party projects can include the caluma urls like this:

```python
from django.conf.urls import include, url

urlpatterns = [
    # ...
    url(r"", include("caluma.caluma_core.urls")),
    # ...
]
```