fix(models): slug fields should be 127 chars
Slug field length is limited to 127 chars, so we can use combined
slugs in many-to-many rel tables as a natural primary key (for example
form-slug.question-slug in the FormQuestion model). The NaturalKeyModel
which we built for this has a CharField primary key with max length of
255 chars, which is exactly two 127-char slugs and a dot inbetween.

Note that since for some reason the foreign keys' types sometimes deviate
from their target fields, we need a way to ensure this won't happen,
as well as a way to fix it.

Why this happened in the first place is a mystery that I couldn't figure
out.  It wasn't reproducible in a standalone reproduction project (no
caluma), and I couldn't figure out what part in Caluma's model would cause
it.  Either way, we do need the code to fix this issue, as the broken code
(and with it, some broken database schema) is already out there.