feat(validation): cleanup context data for jexl expressions
Removed superfluous data from JEXL context, added structure explicitly as
"info" object. This should be easier to access and use.
Additionally, there now is also a `root` object within the `info` that can
equally be used to access the top level form. This may in the future be the
primary way to go, allowing us to deprecate the `form` variable.