feat(workflow): "controlling" for Tasks and WorkItems
Analogous to `address_groups`/`addressed_groups`, this commit implements a
`control_groups` JEXL on `Task` that evaluates to a list of
`controlling_groups` on `WorkItem`.

For this, also the evaluation context for the JEXL evaluation is
expanded.

\## Extending the JEXL context

This commit adds the following new variables:

 - `work_item_creating_group`
 - `case_creating_group`
 - `prev_work_item_controlling_groups`

\### `work_item_creating_group`

This is useful to assign the group that started a work_item as controllers.

\### `case_creating_group`

Same as above, but use the group that started the case.

\### `prev_work_item_controlling_groups`

Let's say you have a work_item "A", that has group "G" assigned as controlling
group.
When the work_item is finished, you want to start a "controlling"
work_item "B" addressed to the controlling group "G" of the work_item "A".

This can be achieved if you set `address_groups` of the `Task` for
work_item "B" to `work_item_controlling_groups`. This will then correctly
evaluate to group "G".

Closes #969