fix(forms): ensure jexl evaluation is local to the relevant question
When evaluating a JEXL expression (is_required / is_hidden in questions),
we need to make sure that the context is set to the correct form / field set.
Otherwise, we may end up with cross talk between table rows, for example,
as the answer transform would just fetch the first answer, which may be from
another row than what we're currently looking at.