feat(arch): move business logic for case creation to custom manager
When caluma is used as a django app, it is often desireable to trigger
specific mutations by calling python APIs directly (instead of the
GraphQL API). This shows how we could go about offering a Python API for
starting cases that mirrors what the GraphQL API offers via saveCase: We
extract the business logic out of the serializer into a custom manager.