feat(architecture): expose start case logic in public python api module
When caluma is used as a django app, it is often desireable to trigger
specific mutations by calling python APIs directly (instead of the
GraphQL API). This shows how we could go about offering a Python API for
starting cases that mirrors what the GraphQL API offers via saveCase: We
extract the business logic out of the serializer into a domain logic
class which is then used in a public method exposed in an api module.