feat(management): extend cleanup-command for cleaning up dangling models
Add option -d, --dangling, cleaning up instances with missing relation.
For example, if a question was deleted, the historical answers
referencing this question might have a broken reference.
Passing "-d" will include theses instances for deletion.

Caution: If this option is provided, the --keep filter is not respected
for such instances!