refactor: fields list should be list, not tuple
The DRF documentation shows the fields list to be a list, not a
tuple. When analyzing the code using MyPy, it assumes that an N-tuple
is sort of a type, and changing the type to an N+X-tuple in a subclass
is therefore a type validation (which I sort of agree with).

Therefore, let's use lists instead.