feat(default_answers): handle table_questions
This commit handles default_answers on table_questions. If a table
question has a default_answer for a row_question that has a different
default_answer, the default_answer on the default table_answer wins. If
there is no default in the row, the default_answer of the row_question
is used.