feat(search): support multiple choice and improve choice search
Add support for searching multiple choice questions (both dynamic and
static). And while we're at it, also change the lookup of choice value
to use the option's label instead of the slug, which ought to be much
more useful.

Note: This implies that the label is present in the user's current
language, otherwise it won't find possible matches.