feat: calculated float question
Support calculated question of type float.
This commit adds all necessary bits for calculated questions with the
implementation of the calculated_float type. Answers for a calculated
question will be calculated from its `calcExpression` jexl.

Example:
slug: "total-price"
calcExpression:
'"quantity"|answer * ("product-price"|answer * (1 + "vat"|answer))'

Caluma will try to generate the answers as soon as possible, with a
default value of `None`. The value will be updated whenever the
expression is valid (eg. all referenced answers are filled).
Recalculation occurs whenver a referenced answer is updated or if the
calcExpression is changed.

Limitations:
Referencing another calculated question is not permitted and will yield
a validation error.