fix: extract_global_id should not be too eager in decoding
We need to validate the extracted type in `extract_global_id()`, as it
may lead to errors when passing a valid slug, that's also valid base64,
and happens to contain a colon.

Instead of naively decoding and returning an invalid (type, identifier)
pair, validate the type's name and behave correctly in those cases.