feat(minio): allow disabling ssl verification (#1408)
When running Caluma in a local dev environment that uses HTTPS (for
example, using the Caddy web server), working with Minio wouldn't work
anymore, as Python (and in turn, the Minio client) would check the
certificate's validity.

Introduce a new setting MINIO_DISABLE_CERT_CHECKS to bypass this in
development environments.  Explicitly verify that DEBUG is set to True
as well, as we don't want this to be used in production.