feat(filters): clean up all the old filter syntax
BREAKING CHANGE:

We have for a long time had two syntaxes for filtering, one of which was
deprecated for over a year now and only kept for backwards
compatibility.  Finally, the old form has to go.

Old:
```graphql
query foo {
    allWorkItems(status: $status) {...}
}
```

New:
```graphql
query foo {
    allWorkItems(filter: [{status: $status}]) {...}
}
```

The new syntax also allows reusing the same filter type twice to narrow
down the result set even more. It also supports inverting the filters to
implement explicit exclusion. It is now time to get rid of the old
syntax.

The following changes are breaking:

- Removed deprecated `slug` filters in favor or `slugs` filters
- Removed top level filters
- Removed `orderByQuestionAnswerValue` filter in favor of
`documentAnswer` in the top level orderset
- Removed `orderBy` in filtersets in favor of top level `order` orderset
- Removed ordering by static meta attributes (configured with
`META_FIELDS` in favor of `meta` in the top level orderset

Resolves #1489.
Closes #1490