fix: extend uwsgi buffer
UWSGI has a buffer of 4096 bytes by default, in which all request headers
must fit. If a user has particularly many roles/groups in their JWT token,
this limit is exceeded. UWSGI will then abort the request, and the reverse
proxy in front will respond with an error such as HTTP/502.