feat: introduce analytics - data extraction
The analytics table provides insight into Caluma's data by collecting
information from forms, cases, and so on.

In this version, we can start extracting data from cases and the data that
is attached to them - documents, work items, and their documents in turn,
as well as the answers within the documents.

There are (pre-configured) filters that can limit the output of a table
to certain conditions. Simple value comparison for now, no complex filtering.
This will need to be added at a later stage when we get a bit of experience
with the whole thing.

A commandline tool is added to run the analytics from the shell. Helps with
debugging and testing, but might also be useful fro cron jobs etc.

The analytics can be set to respect the configured visibilities of the data
that is extracted: I enabled, users of the analytics view will not see data
they would not see otherwise either.