Graphene django v3 (#1647)
* fix(tests): adjust test to new error message style

graphene 3.0 returns a text fragment of the offending query

* fix(middleware): use new visitor and ast conventions

The node types are not CamelCase named, but snake_cased. Also,
the operation is an AST object, not a bare string.

In queries without variables, the variables MUST NOT be set to an empty
string. (None works however...)

* fix(cases): various fixes in csae filter and tests

* fix(schema): make schema tests work again (sort of)

The schema test relies on stringification of the root schema object.
This currently fails due to a function within the graphql core package
being unable to handle django lazy translation objects.

The test setup is rigged such that once upstream fixes the problem, it
will trigger a failure here and we can remove the hacky workaround.

* fix(tests): more tests fixed with correct (non-uuid) input objects

Also remove passing some info objects to schema_executor(), it doesn't
accept them anymore (just pass context_value as that's the interesting
bit)

* fix(tests): stringify localized fields when passing to gql

* chore(coverage): improve coverage

Most of the lines that were "not covered" will never happen or are error
cases such as raising validation errors, or other non-problematic parts.