feat(analytics): introduce analytics table (#1572)
* chore: bump python to 3.8

This is needed for analytics, as we need some advanced caching functionality
to build our queries.

BREAKING CHANGE: If you run Caluma as a Python module and still run 3.7,
this will break

* chore: up max complexity a bit

We have some complex tests that now seem to fail (Despite not
changing in this PR)

* refactor(tests): enable form_and_document to be called multiple times

The fixture builds a form and document structure. If we need multiple
documents of the same structure, we should be able to call the fixture
multiple times.

* feat: introduce analytics - data extraction

The analytics table provides insight into Caluma's data by collecting
information from forms, cases, and so on.

In this version, we can start extracting data from cases and the data that
is attached to them - documents, work items, and their documents in turn,
as well as the answers within the documents.

There are (pre-configured) filters that can limit the output of a table
to certain conditions. Simple value comparison for now, no complex filtering.
This will need to be added at a later stage when we get a bit of experience
with the whole thing.

A commandline tool is added to run the analytics from the shell. Helps with
debugging and testing, but might also be useful fro cron jobs etc.

The analytics can be set to respect the configured visibilities of the data
that is extracted: I enabled, users of the analytics view will not see data
they would not see otherwise either.