feat(analytics): pivot table implementation
The pivot table implements aggregate queries for the analytics module. It
allows taking data from the SimpleTable, and aggregating and grouping
the result.

The implementation deviates a bit from the spec: We originally intended to
use different fields for the "grouping" and "aggregating" parts. However
it was realized that those differ in fact only in the way they're
displayed, not in the way they need to be defined in the GQL level. So
now, analytics tables only have one type of fields instead of two.

Instead of having "row fields" and "data fields", the aggregate fields
now have a pseudo-function "GROUP" which marks a field as "GROUP BY"
instead of an aggregate function. Those will then need to be rendered
as rows in the (frontend) output.

Also, instead of having distinct "extraction" and "pivot" tables that
depend on each other, we now only have one type of table. If you choose
to apply aggregate functions to a field, it will be a pivot table, otherwise
it will be a plain "extraction" table.