feat(test): add at_date fixture
The at_date fixture can run a piece of code (specified as a
callback/lambda) at a given fixed date. This is useful for calling
factories, as Django normally wouldn't allow overriding `auto_add_now`
date fields that we use for `created_at` attributes.

So, this enables us to create objects at different points in time,
disregarding the current time (even if frozen)