feat(core): allow input type override
The original Graphene API allows adding an `Input` (or, without Relay, a
`Arguments`) class to be added on mutation classes. Caluma derives it's
inputs solely from the attached serializer.

This fails if the required data type is more complex than what can
be represented on the serializer, whose most complex type is a list
of strings.

However, we need a list of well-defined objects, so we check on the
mutation class to see if a more specific field type is declared, and
use that one in the schema.