fix(log): minio stat errors shouldn't be errors
When an object in Minio cannot be `stat`ed, this should not be logged
as an error: It may be quite a normal result, for example if a file has
not been uploaded yet, and Caluma for some reason wants to check it.
Therefore, degrading the log message from "error" to "warning".

Also, for debugging reasons, we're adding the object name to the
log message: Without it, we're getting errors in the log and have no
chance of figuring out why it happened.