feat(analytics): add support for form name extraction
While at it, for better "generic" extraction, we now also allow extracting
answers from documents without specifying the form in the data source,
using `document[*].answer_slug` syntax.

Note: This is kind-of breaking, because we replaced the `form_id` field
on documents by `caluma_form.slug`. However, it's not "officially" breaking
as the analytics part is still considered unstable.