fix(analytics): do not crash with more complex visibility layers
Analytics uses generated SQL from the visibility layer of the application.
However, only the part after the WHERE is actually used. Now if a
visibility layer adds expressions that affect the FROM bit
of an SQL statement, and these expressions contain query parameters,
we have too many parameters in for the remaining SQL fragment.

Those parameters need to be discarded, as we don't use the SQL that
came along with them.