feat: allow using client id as user for client-secret auth
When authentication happens with a "service user" instead of
a regular user, we may not get a username. However, we can now
configure Caluma to use the client ID as username instead.

This is optional to ensure that the behaviour is intended, and
no "wrong" fallback happens in case the username claim was just
configured incorrectly.