feat(auth)!: only use userinfo for auth (drop introspect)
For OIDC-auth it's not necessary to have a fallback to the introspect
endpoint, so this commit drops it. For this to work, make sure to include the
`openid`-scope in the request.

BREAKING CHANGE: setting the `OIDC_INTROSPECT_*` and
`OIDC_CLIENT_AS_USERNAME` env vars is obsolete now. Auth now only calls the
userinfo endpoint without any fallback to the introspect endpoint.
This will possibly have an impact on the `username` property of the user
object. If your extensions depend on that property, please update them
if necessary.
Additionally, all requests sent to Caluma must include the
`openid`-scope.