fix(healthz): do not warn about expected events
As part of the health-check routines, a file is removed in minio and
afterwards it's asserted that the file is gone. This call to
`Minio.stat_object()` led to a warning being issued during every
health-check iteration like the following:

```
Minio error, cannot stat object 'healthz_tmp-object_{UUID}': NoSuchKey
```

This commit extends `Minio.stat_object()`, so that issuing of the
log-warning can be suppressed by profiding the argument
`suppress_warning=True`. Default behaviour doesn't change and will still
issue this warning.