feat(copy): adds on_copy to datasources to alter dynamic values
Add a new method on_copy on datasources. When a an answer of type
TYPE_DYNAMIC_CHOICE or TYPE_DYNAMIC_MULTIPLE_CHOICE gets copied the
meaning of the slug and label of the dynamic option could potentially
be changed when the datasource it's data changes. During the answer copy process
the linked datasource it's on_copy method will be called to decide the
desired result.

The new method takes the old answer, the new answer and a slug,label tuple of the existing dynamic
option as arguments, and can:

- return the same slug,label tuple to not perform any change (default behavior)
- return an altered slug or(/and) label to change the answer value
- return a None value for the slug, to discard the answer value