feat(structure): introduce a fast-loading mechanism
Introduce a FastLoader class that is able to preload a full
document/form structure into memory, with as few and simple queries
as possible.

This reduces the number of DB hits the code needs to perform during
document validation.

Note some tests had to be fixed as well - adding family attributes
to some row documents, as the fast-loader is even more picky than the
structure code was about it.