fix(jexl): fix crashes in form validation when validating options
The new structure code has a strict policy of only doing weakrefs from
child fields to their parents, to avoid circular dependencies and therefore
causing GC leaks.

However the validation code had a situation where a child field was used
while the (root) field went out of scope, therefore losing the form structure.

This is now fixed - the private utility method `_structure_field()`
now returns the root field as well, even though it's not actually being used.

While we're at it, also raise a ConfigurationError if a field is missing in
a form.