fix(form): error when saving an answer that does't belong to a form
When saving an answer, we save it to the DB; however if it has calculation
dependents, we load the full structure to recalculate these answers.

This allows us to detect missing questions in the form structure. As this is
a violation of our assumptions, we now throw an explicit exception