fix(history): make history aware of potential proxy models
`django-simple-history` currently creates historical records for proxy
models as well instead of using the historical records of the base
model. This issue is tracked in jazzband/django-simple-history#544 but
doesn't seem to get any attention currently which is why we applied the
workaround from this comment:

https://github.com/jazzband/django-simple-history/issues/544#issuecomment-1538615799

This will make sure that changes on a proxy model will be tracked in the
historical records of the base model.