fix(copy); change order of document-answer on_copy
Instead of updating dynamic options during copying answers, we now wait for all
documents and subdocuments to be copied before updating the dynamic options. This
ensures that all answers that might be linked to dynamic options are copied and
available to link the new dynamic options to the copied document.