fix(jexl): remove whitespace in stringify transform
Remove whitespace in the result of stringify transform to ensure that the
result of the stringify transform in caluma corresponds to the stringify
transform in ember-caluma. The JSON.stringify method on an array doesn't
include any whitespaces, while json.dumps of an array includes whitespaces
between the elements by default, which leads to issues when comparing
the result with values in a jexl. This can be set through the separators
parameter in the json.dumps method.